pyproject.toml
[project]
name = "agb-database-with-sqlalchemy"
version = "0.1.0"
description = ""
authors = [
    {name = "a.dolgov",email = "dolgov_am@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.37,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "orjson (>=3.10.15,<4.0.0)",
    "authx (>=1.4.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

